requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
python-jose
werkzeug
passlib
bcrypt
jwtErrors
